feat: :recycle: Refactored type declarations
Separated type declaration file into multiple files to improve separation of concerns and have type declarations associated with each specific response.
