+++ src/models/comic-chapters.d.ts
@@ -1,0 +1,7 @@
+import { Chapter } from "./common-types";
+
+export interface ComicChapters {
+  chapters: Chapter[];
+  total: number;
+  limit: number;
+}
